+++ src/main/java/insane96mcp/enhancedai/modules/mobs/anticheese/AntiCheese.java
@@ -1,11 +1,16 @@
-package insane96mcp.enhancedai.modules.base.anticheese;
+package insane96mcp.enhancedai.modules.mobs.anticheese;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.monster.Enemy;
 import net.minecraft.world.entity.vehicle.Boat;
@@ -14,9 +19,10 @@ import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.EntityMountEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Anti-Cheese", description = "Prevent players from abusing some game mechanics to stop mobs")
-@LoadFeature(module = Modules.Ids.BASE)
+@Label(name = "Anti-Cheese", description = "Prevent players from abusing some game mechanics to stop mobs. Mobs in the entity type tag enhancedai:ignore_anti_cheese will not be affected by this feature.")
+@LoadFeature(module = Modules.Ids.MOBS)
 public class AntiCheese extends Feature {
+    public static final TagKey<EntityType<?>> IGNORE_ANTI_CHEESE = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "ignore_anti_cheese"));
 
     @Config
     @Label(name = "Prevent Boating & Minecarting", description = "If true, 'Enemies' will no longer be able to be Boated and Minecarted.")
@@ -34,7 +40,8 @@ public class AntiCheese extends Feature {
     public void onMount(EntityMountEvent event) {
         if (!this.isEnabled()
                 || !(event.getEntityMounting() instanceof Enemy)
-                || !(event.getEntityBeingMounted() instanceof Boat) && !(event.getEntityBeingMounted() instanceof Minecart))
+                || !(event.getEntityBeingMounted() instanceof Boat) && !(event.getEntityBeingMounted() instanceof Minecart)
+                || event.getEntityMounting().getType().is(IGNORE_ANTI_CHEESE))
             return;
 
         if (preventBoating)
@@ -44,7 +51,8 @@ public class AntiCheese extends Feature {
     @SubscribeEvent
     public void onJoinLevel(EntityJoinLevelEvent event) {
         if (!this.isEnabled()
-                || !(event.getEntity() instanceof Mob mob))
+                || !(event.getEntity() instanceof Mob mob)
+                || mob.getType().is(IGNORE_ANTI_CHEESE))
             return;
 
         if (antiBoatAndMinecart)